Fix reassessment flow
This is mostly reverts d1e61aef6e250b138715de844945c77fbd1f3b11
and d1e61aef6e250b138715de844945c77fbd1f3b11 which removed the
conditional for the link "Yes, they are correct" if it's a
reassessment which changed the flow. It keeps the change to
the breadcrumbs by renaming it to "Check details".