Fix foreign key contraint in approval API endpoint
Before in manager approval, we'd also assign a panel approver
but now we’ve added the ability to skip it so we’re setting
`assignedTo` to an empty string which violates the foreign key
constraint of it being a user’s email address.

This updates it so we only update `assignedTo` if it's the action
is needs panel approval i.e. QAM.