Update workflow panel meta text
For assignments, only shows when the assignee is not the current user
and if it's an in-progress workflow it will show if the person who
created it is not assigned to it.

Now shows the status action e.g. "Submitted by Jane Doe on 4 Aug 2021"
if not the current user, otherwise "Submitted by me on 4 Aug 2021".

Previously, we didn't show the date or always show the status action.