Refactor index page to use new useQueryParams hook
This fixes an issue when a user is on All tab with pagination and then
switches to Work assigned to me and ends up showing no results. It was
due to not the query params in the URL and in state not being in sync
as well as not being able to update multiple at the same time.