Fix inability to sign in when a historic user
We've auto-created users when importing historic workflows but this has
caused an issue when trying to sign in between Next Auth can't auto-link
an account with the matching the email address of an existing user.

As a result, we've added in the sign in callback to create an account
for the user, update name and image as we didn't have that information
when importing.

Co-authored-by: Jaye Hackett <jhackett1@users.noreply.github.com>