Fix review status for workflows without a need for panel approval
For workflows that didn't require QAM or panel approval, it returned
no action. This is because previously, you *had* to go through through
panel so the code relied on panelApprovedAt. Therefore, the fix is to
make sure we consider workflows that have been manager approved and don't
need panel approval.