fix `node-fetch` vulnerability
Running `npm audit` raised a vulnerability issue with `node-fetch` and recommended that we install `next@12.0.10`  to resolve the vulnerability.

However, after testing the nextJS upgrade in the pipeline, the browser tests failed and threw type errors: `res.once is not a function`.
These seemed to be coming from a file in the `@sentry/nextjs` package.

Running `npm update` to update all the packages, updated `@sentry/nextjs` from v6.15.0 to v^6.17.4.

After doing this the browser tests didn't fail, so hopefully that fixed the issue.