Fix `node-fetch` vulnerability by running npm update
Updating the entire version of NextJS seemed to introduce some issues with content security
and the use of eval in files in the next node module.

We've reverted that change and it looks like we can also get rid of the vulnerability if we run
`npm update`. So trying that approach to see if this works instead.