Keep pipeline simple and linear
Grouping them as 'content changes' and running all the imports and pushes to S3 regardless of which update was done in the pipeline.
This way, we always get the most recently published version for each and not updates isolated to specific pipeline runs.
This should  reduce the risk of deploying stuff in the wrong order or deploying a pipeline with a missing change.